Refactor normal and vision loaders
